requirements.txt
flask==3.1.0
twilio==9.4.0
cryptography==44.0.0
python-dotenv==1.0.1
